Extract presence-penalty helpers into a light leaf module so the test collects
The backend-CI matrix installs torch and transformers but not
unsloth/peft, and test_presence_penalty.py imported through inference.py,
whose module-scope from unsloth import ... raised PackageNotFoundError and
aborted collection for the whole session. Move the pure torch+transformers
helpers (apply_presence_penalty, _make_presence_penalty_processor) into
core/inference/presence_penalty.py; inference.py re-imports them for its
runtime use and the test imports the leaf directly, so the unit tests run
in the lightweight env. Also run the PR's MLX additions through the repo
formatter (pre-commit.ci).
